FINI

+++ Assets/Scripts/Inn/Inn.cs
@@ -13,27 +13,26 @@ public class Inn : MonoBehaviour
         {
             if (PlayerPrefs.GetInt("classe") == 0)
             {
-                PlayerPrefs.SetInt("Santé", 20);
+                PlayerPrefs.SetFloat("Santé", 20);
             }
             else if (PlayerPrefs.GetInt("classe") == 1)
             {
-                PlayerPrefs.SetInt("Santé", 15);
+                PlayerPrefs.SetFloat("Santé", 15);
             }
             else
             {
-                PlayerPrefs.SetInt("Santé", 10);
+                PlayerPrefs.SetFloat("Santé", 10);
             }
 
-            PlayerPrefs.SetInt("Santé_AI1", 20);
-            PlayerPrefs.SetInt("Santé_AI2", 15);
-            PlayerPrefs.SetInt("Santé_AI3", 10);
+            PlayerPrefs.SetFloat("Santé_AI1", 20);
+            PlayerPrefs.SetFloat("Santé_AI2", 15);
+            PlayerPrefs.SetFloat("Santé_AI3", 10);
         }
     }
 
     bool EnoughMoney()
     {
         int Gold = ListeLoadGold();
-
         if (Gold < 10)
         {
             this.LOG("Vous n'avez pas assez d'argents pour aller à l'auberge");
